add net API paddle.tensor.matmul
