+++ python/paddle/tensor/__init__.py
@@ -15,3 +15,4 @@
 # TODO: import all tensor api under this directory, 
 # including tensor creation, manipulation, math, etc.
 # __all__ = [ ]
+from paddle.tensor.linalg import *
